[fix] split functions into seperate pages
