+++ client/components/parents/parentsPage.js
@@ -1,5 +1,4 @@
 import { createAnElement } from "../../utils/elementCreator.js";
-import { renderHeader } from "../header.js";
 import { renderTally } from "./tally.js";
 //import { taskReport } from "./taskReport.js";
 import { tasksList } from "./tasksList.js";
@@ -12,7 +11,10 @@ export async function renderParentsPage() {
   // const tallySection = await renderTally();
   // const taskReportSection = await taskRepost();
   // const tasksSection = await tasksList();
-  const [tallySection, tasksSection] = await Promise.all([renderTally(), tasksList()]);
+  const [tallySection, tasksSection] = await Promise.all([
+    renderTally(),
+    tasksList(),
+  ]);
 
   const parentsPageWrapper = createAnElement(
     "div",
